Clean up library structure
